refactor(perl-corpus): extract metadata and loading facade modules (#0000)
