+++ crates/perl-corpus/src/metadata/mod.rs
@@ -1,0 +1,7 @@
+mod ids;
+pub mod parser;
+mod query;
+mod section;
+
+pub use query::{find_by_flag, find_by_tag};
+pub use section::Section;
