+++ crates/perl-corpus/src/prelude.rs
@@ -1,0 +1,1 @@
+pub use crate::{CorpusPaths, Section, find_by_flag, find_by_tag, parse_dir, parse_file};
